Add support for escaping end of line with '\\'
Ends of lines can now be escaped with backslashes à la bash so:

    hello = world\
    !

is interpreted as a single line: "hello = world!".

Note: Escaping backslashes does not work so

    hello = world\\
    !

is interpreted as: "hello = world\!".

Note2: Inline bash lines are not subject to this end-of-line escaping,
so:

    :echo o\
    :o

will still result in the expected shell expression, "echo oo", and not
"echo o:o", which would be confusing.