Fix goto_subdir shortcuts not working with spaces
The macros defined in presets/defaults using <change-folder> were not
working when paths containing spaces were provided. This is due to the
fact that spaces are interpreted specially by mutt in this context. To
fix that, all spaces are now escaped by <quote-char> for these commands.

Original bug report by Trannie Carter:

     Now that I have mutt.record pointing to `[GMail].Sent Mail` and
     have tried to use the `_s` macro, I've discovered that mutt's
     interactive <change-folder> behavior doesn't play nice with a macro
     playback that contains spaces.

     Normally when you interactively change folders, mutt responds
     immediately to a space character.  It seems to reset the path
     specification.