Fix 'mailboxes' being statically evaluated
In presets/defaults, the 'mailboxes' variable was being statically
evaluated. That is, the list of folders was being generated by mutterer
itself. This is undesirable since it forces the user to run 'mutterer
install' every time a new folder is added, which can cause some
confusion as well.

Now, mutt evaluates the bash expression used to find all the boxes
itself.