Add support for matching [tags] in 'lists' preset
The 'lists' preset used to only match against email addresses using a %L
pattern to determine if an email belonged to a given mailing list. It
is now also capable of matching by the tags that appear at the beginning
of subject lines in some mailing lists (e.g. '[Haskell-cafe]').