Make sync shortcut invoke offlineimap with '-o'
If offlineimap was configured with the autorefresh option, in mutt, the
sync shortcut (default 'S') would spawn an instance of offlineimap which
would never exit. This patch adds the '-o' option to this invocation of
offlineimap in order to make it ignore the autorefresh setting and exit
immediately after syncing.