[refac+bugfix] presets: Avoid modifying IFS
IFS can have effects deep in the shell stack and is annoying to keep
track of, so we're better off using alternatives to `IFS=$'\n'`.

Note: This commit actually fixes a latent bug in shell function
'all_accounts_except' whereby if one of the provided exceptions was not
an actual account, it would appear in the output along with the filtered
accounts.